refactor: :recycle: move main entity types to own file
remove routes
