+++ src/main-entity-types.ts
@@ -1,0 +1,34 @@
+export type MainEntityType =
+	| 'Article'
+	| 'Blog'
+	| 'BlogPosting'
+	| 'Book'
+	| 'Chapter'
+	| 'Code'
+	| 'CreativeWork'
+	| 'Dataset'
+	| 'Event'
+	| 'FAQPage'
+	| 'HowTo'
+	| 'HowToSection'
+	| 'HowToStep'
+	| 'ImageObject'
+	| 'ItemList'
+	| 'JobPosting'
+	| 'LocalBusiness'
+	| 'Movie'
+	| 'MusicPlaylist'
+	| 'MusicRecording'
+	| 'NewsArticle'
+	| 'Organization'
+	| 'Person'
+	| 'Product'
+	| 'Question'
+	| 'Recipe'
+	| 'Report'
+	| 'Review'
+	| 'ScholarlyArticle'
+	| 'SoftwareApplication'
+	| 'TechArticle'
+	| 'VideoObject'
+	| 'WebPage'
