--- src/routes/scholarly-article/+page.svelte
@@ -1,15 +1,0 @@
-<script lang="ts">
-	import Head from '$lib/components/head.svelte'
-</script>
-
-<Head
-	url="https://example.com/scholarly-article"
-	title="Scholarly Article"
-	description="This is a sample scholarly article."
-	contentType="ScholarlyArticle"
-	authorName="Dr. Jane Smith"
-	authorUrl="https://example.com/authors/dr-jane-smith"
-	image="https://example.com/images/scholarly-article-image.jpg"
-	datePublished="2023-04-05T10:00:00Z"
-	dateModified="2023-04-05T12:00:00Z"
-/>
