--- src/routes/tech-article/+page.svelte
@@ -1,15 +1,0 @@
-<script lang="ts">
-	import Head from '$lib/components/head.svelte'
-</script>
-
-<Head
-	url="https://example.com/tech-article"
-	title="Tech Article"
-	description="This is a sample tech article."
-	contentType="TechArticle"
-	authorName="John Techwriter"
-	authorUrl="https://example.com/authors/john-techwriter"
-	image="https://example.com/images/tech-article-image.jpg"
-	datePublished="2023-04-05T10:00:00Z"
-	dateModified="2023-04-05T12:00:00Z"
-/>
